src/test/java/com/example/ecommerceplatform/service/PayPalServiceTest.java
package com.example.ecommerceplatform.service;

import com.paypal.api.payments.*;
import com.paypal.base.rest.APIContext;
import com.paypal.base.rest.PayPalRESTException;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Disabled;
import org.junit.jupiter.api.Test;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;

import static org.mockito.Mockito.*;
import static org.junit.jupiter.api.Assertions.*;

public class PayPalServiceTest {

    @InjectMocks
    private PayPalService payPalService;

    @Mock
    private APIContext apiContext;

    @Mock
    private Payment payment;

    @BeforeEach
    public void setup() {
        MockitoAnnotations.initMocks(this);
    }

    @Test
    @Disabled
    public void testCreatePayment() throws PayPalRESTException {
        // Arrange
        String amount = "100.00";
        String currency = "USD";
        String method = "paypal";
        String intent = "sale";
        when(apiContext.getClientID()).thenReturn("your_client_id");
        when(apiContext.getClientSecret()).thenReturn("your_client_secret");

        // Mock the Payment.create() method to return a mocked Payment object
        when(payment.create(apiContext)).thenReturn(payment);

        // Act
        Payment result = payPalService.createPayment(amount, currency, method, intent);

        // Assert
        assertNotNull(result);
        // Add more assertions as needed
    }

    @Test
    @Disabled
    public void testExecutePayment() throws PayPalRESTException {
        // Arrange
        String paymentId = "your_payment_id";
        String payerId = "your_payer_id";
        when(apiContext.getClientID()).thenReturn("your_client_id");
        when(apiContext.getClientSecret()).thenReturn("your_client_secret");

        // Mock the Payment.execute() method to return a mocked Payment object
        when(payment.execute(apiContext, any(PaymentExecution.class))).thenReturn(payment);

        // Act
        Payment result = payPalService.executePayment(paymentId, payerId);

        // Assert
        assertNotNull(result);
        // Add more assertions as needed
    }
}